+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## {{ UNRELEASED_VERSION }} - [{{ UNRELEASED_DATE }}]({{ UNRELEASED_LINK }})
 
+## v30393237885.70.1-build.6b86d9fcd4b364435a4e81c5f4563513048756d1 - [July 28, 2026](git://github.com/tanaabased/prepare-release-action.git)
+
 ## v30390649808.69.1-build.def88140ca1c4ae0b03685b0dba1f465d8ee7b05 - [July 28, 2026](git://github.com/tanaabased/prepare-release-action.git)
 
 ## v30389116470.68.1-build.6a13252a4281bb00cbd64ce68f7d334df8923f04 - [July 28, 2026](git://github.com/tanaabased/prepare-release-action.git)